Update form layout
